refactor: rearrange page files into separate folders
