+++ src/pages/MarketHistory/index.js
@@ -1,8 +1,8 @@
 import React, { useState } from "react";
 import Styled from "styled-components";
-import SellHistoryContainer from "../components/mypage/market/SellHistoryContainer";
-import PurchaseHistoryContainer from "../components/mypage/market/PurchaseHistoryContainer";
-import Navbar from "../components/Navbar";
+import SellHistoryContainer from "../../components/mypage/market/SellHistoryContainer";
+import PurchaseHistoryContainer from "../../components/mypage/market/PurchaseHistoryContainer";
+import Navbar from "../../components/Navbar";
 
 const MarketHistoryWrapper = Styled.div`
   margin: 110px;
